refactor ingest_data.py
